Use ViperConfig in LanguageServerRunner
